requirements.txt
django
gunicorn
python-dotenv
whitenoise
django-admin-corporate==1.0.1
python-fsutil
whitenoise
gunicorn
dj-database-url
psycopg2-binary
django_custom_error_views
# mysqlclient
google-generativeai